packages/dbgear-editor/pyproject.toml
[project]
name = "dbgear-editor"
version = "0.1.0"
description = "FastHTML-based web editor for DBGear schema and data management"
authors = [
    {name = "tamuto", email = "mutomob@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-fasthtml>=0.12.24,<0.13.0",
    "monsterui>=1.0.23,<2.0.0",
    "uvicorn>=0.30.0",
    "dbgear"
]

[project.scripts]
dbgear-editor = "dbgear_editor.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
